Add a helper to assist in configuring repositories and remotes for exporting metadata
