+++ GitTfs/GitTfs.csproj
@@ -202,6 +202,7 @@
     <Compile Include="Util\Bouncer.cs" />
     <Compile Include="Util\ChangeSieve.cs" />
     <Compile Include="Util\ConfigPropertyLoader.cs" />
+    <Compile Include="Util\ExportMetadatasInitializer.cs" />
     <Compile Include="Util\PathResolver.cs" />
     <Compile Include="Util\CommitSpecificCheckinOptionsFactory.cs" />
     <Compile Include="Core\GitTfsVersionProvider.cs" />
